requirements.txt
1.8.7 <= django == 1.8.11
python == 2.7.10
pillow == 3.2.0
django-reversion == 1.10.1
pytz == 2016.6
mysql-python == 1.2.5
